requirements.txt
Django==4.2
pillow==10.3.0
python-dotenv==1.0.1
psycopg2==2.9.9
djangorestframework==3.14.0
